Add CourseResource

+++ app/Http/Controllers/CourseController.php
@@ -2,12 +2,13 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Course;
-use Illuminate\Http\Request;
-use App\Services\FileStorageService;
-use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\StoreCourseRequest;
 use App\Http\Requests\UpdateCourseRequest;
+use App\Http\Resources\CourseResource;
+use App\Models\Course;
+use App\Services\FileStorageService;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CourseController extends Controller
 {
@@ -21,9 +22,9 @@ class CourseController extends Controller
      */
     public function index()
     {
-            $courses = Course::all();
+        $courses = Course::all();
 
-            return response()->json($courses, 200);
+        return CourseResource::collection($courses);
     }
 
     /**
@@ -31,65 +32,60 @@ class CourseController extends Controller
      */
     public function store(StoreCourseRequest $request)
     {
-            $validated = $request->validated();
-            $validated['user_id'] = Auth::id();
-            if ($request->hasFile('thumbnail')) {
-                $validated['thumbnail'] = $request->file('thumbnail')->store('courses', 'public');
-            }
-            $course = Course::create($validated);
+        $validated = $request->validated();
+        $validated['user_id'] = Auth::id();
+        if ($request->hasFile('thumbnail')) {
+            $validated['thumbnail'] = $request->file('thumbnail')->store('courses', 'public');
+        }
+        $course = Course::create($validated);
 
-            return response()->json([
-                'status' => '200',
-                'message' => 'Course created successfully',
-                'data' => $course,
-            ]);
+        return response()->json([
+            'status' => '200',
+            'message' => 'Course created successfully',
+            'data' => new CourseResource($course)
+        ]);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(Course $course)
     {
-            $course = Course::findOrFail($id);
-
-            return response()->json($course, '200');
+        return new CourseResource($course);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function updateCourse(UpdateCourseRequest $request, string $id)
+    public function updateCourse(UpdateCourseRequest $request, Course $course)
     {
-            $course = Course::findOrFail($id);
-            $this->authorize('update', $course);
+        $this->authorize('update', $course);
 
-            $validated = $request->validated();
+        $validated = $request->validated();
 
-            if ($request->hasFile('thumbnail')) {
-                $validated['thumbnail'] = $this->fileStorage->replaceCourseThumbnail($course->thumbnail,$request->file('thumbnail'));
-            }
-            $course->update($validated);
+        if ($request->hasFile('thumbnail')) {
+            $validated['thumbnail'] = $this->fileStorage->replaceCourseThumbnail($course->thumbnail, $request->file('thumbnail'));
+        }
+        $course->update($validated);
 
-            return response()->json($course, 200);
+        return new CourseResource($course);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Course $course)
     {
-            $course = Course::findOrFail($id);
-            $course->delete();
+        $this->authorize('delete', $course);
+        $course->delete();
 
-            return response()->json('Course Deleted successfully', 200); 
+        return response()->json('Course Deleted successfully', 200);
     }
 
-    public function enrollToCourse(Request $request, $courseId)
+    public function enrollToCourse(Request $request, Course $course)
     {
-            $course = Course::findOrFail($courseId);
-            $course->enrolledUsers()->attach(Auth::id());
+        $course->enrolledUsers()->attach(Auth::id());
 
-            return response()->json('User attached successfully', 200);
-        
+        return response()->json('User attached successfully', 200);
     }
 }
